+++ Realm/Realm.Weaver/WeaveResults.cs
@@ -150,16 +150,14 @@ namespace RealmWeaver
             return new WeavePropertyResult(error: error);
         }
 
-        public static WeavePropertyResult Skipped(string reason)
+        public static WeavePropertyResult Skipped()
         {
-            return new WeavePropertyResult(skipReason: reason);
+            return new WeavePropertyResult();
         }
 
         public string? ErrorMessage { get; }
 
         public string? WarningMessage { get; }
-        
-        public string? SkipReason { get; }
 
         [MemberNotNullWhen(true, nameof(Property))]
         public bool Woven { get; }
@@ -183,11 +181,10 @@ namespace RealmWeaver
             Woven = true;
         }
 
-        private WeavePropertyResult(string? error = null, string? warning = null, string? skipReason = null)
+        private WeavePropertyResult(string? error = null, string? warning = null)
         {
             ErrorMessage = error;
             WarningMessage = warning;
-            SkipReason = skipReason;
         }
 
         public override string ToString()